Fix SDPA fusion output layout for GQA prefill (L>1)
The SDPA op always outputs [B,H,L,EV], but GQA attention in Gemma4
produces [B,L,H,EV] after the output reshape-transpose chain. For
decode (L=1) this was a silent no-op; for prefill (L=8, H=8) the
data ordering was wrong, producing garbage logits and empty responses.

Fix: compute a compensating output permutation from the original
reshape-transpose chain and apply it after the SDPA op when it
differs from identity.

Also includes:
- Numerical correctness tests for simple, GQA L=1, and GQA L=8
- _get_scalar() fix in softcap fusion for uniform-valued tensors
- int32 matmul cast fix in decode_coreml.py

+++ gemma_chat/decode_coreml.py
@@ -518,7 +518,10 @@ def chunk_prefill_step(
     write_mask = (jnp.arange(W, dtype=jnp.int32)[:, None]
                   == ring_slots[None, :])  # (W, C)
     any_written = write_mask.any(axis=1)  # (W,)
-    new_ring_vals = jnp.dot(write_mask.astype(jnp.int32), abs_pos)  # (W,)
+    # Use float16 dot — MPS does not support int32 matmul.
+    new_ring_vals = jnp.dot(
+        write_mask.astype(jnp.float16), abs_pos.astype(jnp.float16)
+    ).astype(jnp.int32)  # (W,)
     sliding_pos_ring = jnp.where(any_written[None, :], new_ring_vals[None, :],
                                  sliding_pos_ring)
